Fix: returns

+++ main.py
@@ -1,7 +1,10 @@
-from fastapi import FastAPI, Depends, BackgroundTasks
+from http.client import HTTPException
+
+from fastapi import FastAPI, Depends, BackgroundTasks, HTTPException
+from fastapi.openapi.utils import status_code_ranges
 from sqlalchemy.orm import Session
 
-from schemes import TaskADD
+from schemes import TaskADD, TaskUPDATE
 from database import TaskDB, Base, engine, get_db
 from fastapi.middleware.cors import CORSMiddleware
 
@@ -29,16 +32,13 @@ def send_notification(email : str, background_tasks : BackgroundTasks):
     background_tasks.add_task(write_notification, email, message= 'some notification')
     return {'message' : 'Notification sent in the background'}
 
-@app.get("/")
-def read_root():
-    return {"message" : "Start"}
-
 @app.post("/tasks/")
 def create_task(task : TaskADD, db : Session  = Depends(get_db)):
     new_task = TaskDB(title = task.title, completed = task.completed)
     db.add(new_task)
     db.commit()
-    return {'message' : f'Task {task.id} was created'}
+    db.refresh(new_task)
+    return new_task
 
 @app.get("/tasks/")
 def get_tasks(completed : bool = None, limit : int = None, db : Session = Depends(get_db)):
@@ -52,12 +52,19 @@ def get_tasks(completed : bool = None, limit : int = None, db : Session = Depend
 @app.delete("/tasks/")
 def delete_task(task_id : int, db : Session = Depends(get_db)):
     task = db.query(TaskDB).filter(TaskDB.id == task_id).first()
+    if not task:
+        raise HTTPException(status_code=404, detail=f"Task {task_id} not found")
     db.delete(task)
     db.commit()
-    return {'message' : f'Task {task_id} was deleted'}
+    return task
 
 @app.patch("/tasks/")
-def update_task(task : TaskADD, db : Session = Depends(get_db)):
-    db.query(TaskDB).filter(TaskDB.id == task.id).update({TaskDB.completed : task.completed, TaskDB.title : task.title})
+def update_task(task : TaskUPDATE, db : Session = Depends(get_db)):
+    exist_task = db.query(TaskDB).filter(TaskDB.id == task.id).first()
+    if not exist_task:
+        raise HTTPException(status_code=404, detail=f"Task {task.id} not found")
+    exist_task.title = task.title
+    exist_task.completed = task.completed
     db.commit()
-    return {'message': f'Task {task.id} status was updated'}
+    db.refresh(exist_task)
+    return exist_task
